Silence another race condition (ESRCH instead of ENOENT)
This is one of those things you only observe after running a package like
`proc' from a periodic task (cron job) that runs every minute on a dozen
servers for a couple of weeks :-). The error condition was -correctly-
being swallowed already, but it was more noisy than it needed to be.