Log message when starting GPG agent*
I now understand why the GPG agent hangs in the test suite on Travis CI
(see 526cce7392d) and I intend to resolve this issue today, however this
has highlighted the fact that starting a new GPG agent daemon can have
unintended consequences. As such I think it wise to inform the operator
when this is happening.