Check if gpg-agent is installed (before running it)
The log message added in the previous commit (45b04ed6062) and all of
the associated logic are rather pointless when the gpg-agent program
isn't installed, so let's explicitly check for that!

We don't consider it an error when gpg-agent isn't installed because the
gpg client program should function just fine without the agent running
in the background. If we would raise an exception here then we would
break a workflow that could have worked fine.

We do log a NOTICE message when the gpg-agent program is missing, so
that the operator has a heads up explaining why they may get multiple
password prompts.