Use existing $GPG_AGENT_INFO value when possible*
Recently I upgraded my laptop to Ubuntu 18.04 and noticed that the
with-gpg-agent program is no longer able to find a running GPG agent in
the process tree and so tries to launch a new one, which fails with the
following error message:

  gpg-agent: a gpg-agent is already running - not starting a new one

The silliest aspect of this all is that $GPG_AGENT_INFO is set correctly
already, which makes it inexcusable for with-gpg-agent to fail the way
it does. This is why I've changed the proc.gpg module to check for,
validate and (when possible) re-use existing values of the
$GPG_AGENT_INFO environment variable.

The next commit will tackle the other aspect of the failure I
encountered: The fact that the running GPG agent isn't found.